DEV/marks/requirements.txt
Flask
google-cloud-firestore
google-cloud-aiplatform